+++ app/src/main/java/com/kiral/charityapp/network/services/ProfileService.kt
@@ -1,4 +1,4 @@
-package com.kiral.charityapp.network
+package com.kiral.charityapp.network.services
 
 import com.kiral.charityapp.network.dtos.LoginDto
 import com.kiral.charityapp.network.dtos.ProfileDto
